ENH create a sklearn.utils.multiclass module
