+++ sklearn/multiclass.py
@@ -39,46 +39,6 @@ from .externals.joblib import Parallel
 from .externals.joblib import delayed
 
 
-def unique_labels(*lists_of_labels):
-    """Extract an ordered array of unique labels
-
-    Parameters
-    ----------
-    lists_of_labels : list of labels,
-        Each list of labels have to be in a format supported by
-        ``sklearn.preprocessing.LabelBinarizer``: .
-
-    Returns
-    -------
-    out : numpy array of shape [n_unique_labels]
-        An ordered array of unique labels.
-
-    Examples
-    --------
-    >>> from sklearn.multiclass import unique_labels
-    >>> unique_labels([3, 5, 5, 5, 7, 7])
-    array([3, 5, 7])
-    >>> unique_labels([1, 2, 3, 4], [2, 2, 3, 4])
-    array([1, 2, 3, 4])
-    >>> unique_labels([1, 2, 10], [5, 11])
-    array([ 1,  2,  5, 10, 11])
-    >>> unique_labels(np.array([[0.0, 1.0], [1.0, 1.0]]), np.zeros((2, 2)))
-    array([0, 1])
-    >>> unique_labels([(1, 2), (3,)], [(1, 2), tuple()])
-    array([1, 2, 3])
-
-    """
-    def _unique_labels(labels):
-        lb = LabelBinarizer()
-        lb.fit(labels)
-        return lb.classes_
-
-    if not lists_of_labels:
-        raise ValueError('No list of labels has been passed.')
-
-    return np.unique(np.hstack(_unique_labels(y) for y in lists_of_labels))
-
-
 def _fit_binary(estimator, X, y, classes=None):
     """Fit a single binary estimator."""
     unique_y = np.unique(y)
